[FIX] mail: avoid losing chatter record when changing container
Form view has multiple containers that are exclusive, before this commit
switching from one container to the other would lose all current chatter state.

closes odoo/odoo#102934

X-original-commit: a3d151d0c9c84c2dcffc28a342c40f38d0b28370
Related: odoo/enterprise#32644

+++ addons/mail/static/src/models/chatter.js
@@ -172,6 +172,9 @@ registerModel({
          * @private
          */
         _onThreadIdOrThreadModelChanged() {
+            if (!this.threadModel) {
+                return;
+            }
             if (this.threadId) {
                 if (this.thread && this.thread.isTemporary) {
                     this.thread.delete();
@@ -261,6 +264,9 @@ registerModel({
         }),
         dropZoneView: one('DropZoneView', {
             compute() {
+                if (!this.thread) {
+                    return clear();
+                }
                 if (this.useDragVisibleDropZone.isVisible) {
                     return {};
                 }
@@ -403,6 +409,9 @@ registerModel({
          */
         threadViewer: one('ThreadViewer', {
             compute() {
+                if (!this.thread) {
+                    return clear();
+                }
                 return {
                     hasThreadView: this.hasThreadView,
                     order: 'desc',
@@ -410,10 +419,11 @@ registerModel({
                 };
             },
             inverse: 'chatter',
-            required: true,
         }),
         topbar: one('ChatterTopbar', {
-            default: {},
+            compute() {
+                return this.thread ? {} : clear();
+            },
             inverse: 'chatter',
         }),
         useDragVisibleDropZone: one('UseDragVisibleDropZone', {
